fx_man

+++ source/audiolib/DEBUGIO.C
@@ -20,6 +20,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 #include <stdio.h>
 #include <stdarg.h>
 #include <stdlib.h>
+#include "compat.h"
 #include "debugio.h"
 
 static unsigned short disp_offset = 160 * 24;
@@ -182,7 +183,7 @@ int DB_PrintNum
 
 int DB_PrintUnsigned
    (
-   unsigned long number,
+   uint32_t number,
    int radix
    )
